backend/requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
scikit-learn
pandas
numpy
python-dotenv